fixup(PRVN-003): client-bundle build error — extract pure vocabulary
CI caught a build error local gates miss: the 'use client' liaison
intake form deep-imports SCHOOL_REFERRAL_SERVICES from
school-referral-policy.ts, but the policy module imports @/db/client
(for recordDisclosure) and @sentry/nextjs. Next can't tree-shake those
out of the browser bundle, so the build fails with
'Module not found: Can't resolve fs / net / tls / perf_hooks'.

Same shape as the FND-040b × 'use client' bug captured in STATE.md
known quirks. The feedback memory exists to remind us to run e2e or
next build after barrel/import refactors — this PR's local gates
(biome, tsc, vitest, lint:boundaries) all passed; only `next build`
catches it.

Fix: extract the pure controlled-vocabulary constants
(SCHOOL_REFERRAL_SERVICES, MV_HOUSING_RELATED_SERVICES, type
SchoolReferralService) to school-referral-vocabulary.ts — pure data,
no server-only imports. The policy module still re-exports them for
existing server callers, and now imports them itself for use in
validateMcKinneyVentoBasis. The client form deep-imports from the
new vocabulary file.

Verified locally with `pnpm exec next build`; build now completes.
All other gates remain green: biome, tsc, vitest 360/360,
lint:boundaries.
